Refactor ConsoleInput and Console into a single Console component
