+++ src/main/java/seedu/todo/ui/MainWindow.java
@@ -18,7 +18,7 @@ import seedu.todo.commons.core.Config;
 import seedu.todo.commons.core.GuiSettings;
 import seedu.todo.commons.core.LogsCenter;
 import seedu.todo.commons.events.ui.ExitAppRequestEvent;
-import seedu.todo.ui.components.ConsoleInput;
+import seedu.todo.ui.components.Console;
 import seedu.todo.ui.components.Header;
 import seedu.todo.ui.views.View;
 
@@ -81,10 +81,10 @@ public class MainWindow extends View {
         header.render();
 
         // Load ConsoleInput
-        ConsoleInput consoleInput = ConsoleInput.load(primaryStage, getConsoleInputPlaceholder());
-        consoleInput.consoleOutput = consoleMessage;
-        consoleInput.consoleInputValue = consoleInputValue;
-        consoleInput.render();
+        Console console = Console.load(primaryStage, getConsoleInputPlaceholder());
+        console.consoleOutput = consoleMessage;
+        console.consoleInputValue = consoleInputValue;
+        console.render();
     }
 
     @Override
