+++ src/main/resources/ui/components/Console.fxml
@@ -1,13 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.control.TextField?>
 
-<VBox xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" 
+<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" 
   fx:controller="seedu.todo.ui.components.Console">
   
   <children>
     <TextArea fx:id="consoleTextArea" editable="false" focusTraversable="false" mouseTransparent="true"  
       promptText="Console text" styleClass="console-textarea" wrapText="true" />
+    
+    <TextField fx:id="consoleInputTextField" styleClass="console-input-textfield" onAction="#handleConsoleInputChanged" 
+      onKeyPressed="#handleConsoleInputKeyPress" promptText="Enter command here..." VBox.vgrow="NEVER" />
   </children>
+  
 </VBox>
+
