--- src/main/resources/ui/components/ConsoleInput.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.control.TextField?>
-
-<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" 
-  fx:controller="seedu.todo.ui.components.ConsoleInput">
-  
-  <children>
-    <AnchorPane fx:id="consolePlaceholder" minHeight="80.0" prefHeight="80.0" maxHeight="80.0" VBox.vgrow="NEVER" />
-    
-    <TextField fx:id="consoleInputTextField" styleClass="console-input-textfield" onAction="#handleConsoleInputChanged" 
-      onKeyPressed="#handleConsoleInputKeyPress" promptText="Enter command here..." VBox.vgrow="NEVER" />
-  </children>
-  
-</VBox>
-
